+++ crates/storage/trie_db/redb.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use ethrex_trie::{TrieDB, TrieError};
+use ethrex_trie::{NodeHash, TrieDB, TrieError};
 use redb::{Database, TableDefinition};
 
 const TABLE: TableDefinition<&[u8], &[u8]> = TableDefinition::new("Trie");
@@ -16,7 +16,7 @@ impl RedBTrie {
 }
 
 impl TrieDB for RedBTrie {
-    fn get(&self, key: Vec<u8>) -> Result<Option<Vec<u8>>, TrieError> {
+    fn get(&self, key: NodeHash) -> Result<Option<Vec<u8>>, TrieError> {
         let read_txn = self
             .db
             .begin_read()
@@ -25,12 +25,12 @@ impl TrieDB for RedBTrie {
             .open_table(TABLE)
             .map_err(|e| TrieError::DbError(e.into()))?;
         Ok(table
-            .get(&*key)
+            .get(key.as_ref())
             .map_err(|e| TrieError::DbError(e.into()))?
             .map(|value| value.value().to_vec()))
     }
 
-    fn put(&self, key: Vec<u8>, value: Vec<u8>) -> Result<(), TrieError> {
+    fn put(&self, key: NodeHash, value: Vec<u8>) -> Result<(), TrieError> {
         let write_txn = self
             .db
             .begin_write()
@@ -40,7 +40,7 @@ impl TrieDB for RedBTrie {
                 .open_table(TABLE)
                 .map_err(|e| TrieError::DbError(e.into()))?;
             table
-                .insert(&*key, &*value)
+                .insert(key.as_ref(), &*value)
                 .map_err(|e| TrieError::DbError(e.into()))?;
         }
         write_txn
@@ -50,7 +50,7 @@ impl TrieDB for RedBTrie {
         Ok(())
     }
 
-    fn put_batch(&self, key_values: Vec<(Vec<u8>, Vec<u8>)>) -> Result<(), TrieError> {
+    fn put_batch(&self, key_values: Vec<(NodeHash, Vec<u8>)>) -> Result<(), TrieError> {
         let write_txn = self
             .db
             .begin_write()
@@ -61,7 +61,7 @@ impl TrieDB for RedBTrie {
                 .map_err(|e| TrieError::DbError(e.into()))?;
             for (key, value) in key_values {
                 table
-                    .insert(&*key, &*value)
+                    .insert(key.as_ref(), &*value)
                     .map_err(|e| TrieError::DbError(e.into()))?;
             }
         }